refactor: remove HSR analysis page and update index to support multiple games
- Deleted the HSR analysis page and its associated logic.
- Updated the index page to introduce a new "Supported Games" section.
- Enhanced the features section to reflect the new branding as "GachaScope".
- Improved navigation and statistics calculations to accommodate multiple games.
- Adjusted the UI elements and descriptions for clarity and consistency.
